File upload, show and rename
o	The Admin will Download files from the Input Folder. Rename this folder to Download Forms.
The Admin should not be able to upload anything to this folder.
After the Admin click this folder, he should see a list of all his clients/users. After clicking on a certain client, he should be able to see only the files associated with that user.